🗺️ Atlas: Extract ModelError and remove FailedAttempt duplication
