+++ tests/exit_code_contract.rs
@@ -6,8 +6,9 @@
 
 #![allow(clippy::unwrap_used)]
 
-use maxwells_daemon::error::{ConfigError, EnvError, Error, ModelError};
+use maxwells_daemon::error::{ConfigError, EnvError, Error};
 use maxwells_daemon::exit_code::ExitCode;
+use maxwells_daemon::model::ModelError;
 
 // ── integer code values ───────────────────────────────────────────────────────
